Deprecate passing MARIADB_PASSWORD in favour of mounting a secret
Passing a password in clear text is an obviously bad idea. Add support
for mounting a secret and deprecate the clear text approach.

This change also makes it a fatal error to not pass the password.
Our mariadb-image does not support empty passwords anyway, and we should
not rely on anyone actually using "change_me" without realizing it.

Signed-off-by: Dmitry Tantsur <dtantsur@protonmail.com>